# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Upwork AI Automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all monitored jobs
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [new, analyzing, approved, skipped, applied, rejected]
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Add a job from Chrome extension
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/analyze:
    post:
      operationId: analyzeJob
      tags: [ai]
      summary: Run AI analysis on a job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAnalysis"
  /jobs/{id}/proposal:
    post:
      operationId: generateProposal
      tags: [ai]
      summary: Generate an AI proposal for a job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalGenInput"
      responses:
        "200":
          description: Generated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - in: query
          name: unreadOnly
          schema:
            type: boolean
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, description, budgetType, status, createdAt]
      properties:
        id:
          type: integer
        upworkJobId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        budgetType:
          type: string
          enum: [hourly, fixed]
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        clientCountry:
          type: ["string", "null"]
        clientHireRate:
          type: ["number", "null"]
        clientTotalSpent:
          type: ["number", "null"]
        paymentVerified:
          type: ["boolean", "null"]
        proposalCount:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [new, analyzing, approved, skipped, applied, rejected]
        applyScore:
          type: ["number", "null"]
        riskScore:
          type: ["number", "null"]
        winProbability:
          type: ["number", "null"]
        aiRecommendation:
          type: ["string", "null"]
          enum: [apply, skip, review, null]
        aiReasoning:
          type: ["string", "null"]
        jobUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    JobInput:
      type: object
      required: [title, budgetType]
      properties:
        upworkJobId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        budgetType:
          type: string
          enum: [hourly, fixed]
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        clientCountry:
          type: ["string", "null"]
        clientHireRate:
          type: ["number", "null"]
        clientTotalSpent:
          type: ["number", "null"]
        paymentVerified:
          type: ["boolean", "null"]
        proposalCount:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        jobUrl:
          type: ["string", "null"]
    JobUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, analyzing, approved, skipped, applied, rejected]
        applyScore:
          type: number
        riskScore:
          type: number
        winProbability:
          type: number
        aiRecommendation:
          type: string
          enum: [apply, skip, review]
        aiReasoning:
          type: string
    JobAnalysis:
      type: object
      required: [jobId, applyScore, riskScore, winProbability, recommendation, reasoning]
      properties:
        jobId:
          type: integer
        applyScore:
          type: number
        riskScore:
          type: number
        winProbability:
          type: number
        recommendation:
          type: string
          enum: [apply, skip, review]
        reasoning:
          type: string
        strengths:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string
    Proposal:
      type: object
      required: [id, jobId, content, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: ["string", "null"]
        content:
          type: string
        coverLetter:
          type: ["string", "null"]
        bidAmount:
          type: ["number", "null"]
        estimatedDuration:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, approved, sent, rejected]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProposalGenInput:
      type: object
      properties:
        bidAmount:
          type: number
        tone:
          type: string
          enum: [professional, friendly, confident]
        includePortfolio:
          type: boolean
...
    Notification:
      type: object
      required: [id, type, message, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [high_score_job, message_received, interview_invite, proposal_viewed, contract_offer, info]
        message:
          type: string
        jobId:
          type: ["integer", "null"]
        read:
          type: boolean
        createdAt:
          type: string
...